Move method back to common
