+++ src/common/string/has-template.ts
@@ -1,4 +1,4 @@
-const isTemplateRegex = new RegExp("{%|{{|{#");
+const isTemplateRegex = new RegExp("{%|{{");
 
 export const isTemplate = (value: string): boolean =>
   isTemplateRegex.test(value);
@@ -12,7 +12,7 @@ export const hasTemplate = (value: unknown): boolean => {
   }
   if (typeof value === "object") {
     const values = Array.isArray(value) ? value : Object.values(value!);
-    return values.some((val) => hasTemplate(val));
+    return values.some((val) => val && hasTemplate(val));
   }
   return false;
 };
